Move default attribute registry into object builders
Looks fitting!
