--- fabric-entity-attribute-api-v1/src/main/resources/fabric-entity-attribute-api-v1.mixins.json
@@ -1,12 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.entity.attribute",
-  "compatibilityLevel": "JAVA_8",
-  "mixins": [
-    "DefaultAttributeRegistryAccessor",
-    "DefaultAttributeRegistryMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
