+++ fabric-object-builders-v0/src/main/java/net/fabricmc/fabric/mixin/object/builder/DefaultAttributeRegistryMixin.java
@@ -14,12 +14,18 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.entity.attribute;
+package net.fabricmc.fabric.mixin.object.builder;
 
+import java.util.HashMap;
 import java.util.Map;
 
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
@@ -27,9 +33,14 @@ import net.minecraft.entity.attribute.DefaultAttributeContainer;
 import net.minecraft.entity.attribute.DefaultAttributeRegistry;
 
 @Mixin(DefaultAttributeRegistry.class)
-public interface DefaultAttributeRegistryAccessor {
-	@Accessor("DEFAULT_ATTRIBUTE_REGISTRY")
-	static Map<EntityType<? extends LivingEntity>, DefaultAttributeContainer> getRegistry() {
-		throw new AssertionError("mixin dummy");
+public abstract class DefaultAttributeRegistryMixin {
+	@Shadow
+	@Final
+	@Mutable
+	private static Map<EntityType<? extends LivingEntity>, DefaultAttributeContainer> DEFAULT_ATTRIBUTE_REGISTRY;
+
+	@Inject(method = "<clinit>*", at = @At("TAIL"))
+	private static void injectAttributes(CallbackInfo ci) {
+		DEFAULT_ATTRIBUTE_REGISTRY = new HashMap<>(DEFAULT_ATTRIBUTE_REGISTRY);
 	}
 }
