+++ fabric-object-builders-v0/src/main/resources/fabric-object-builders-v0.mixins.json
@@ -4,6 +4,8 @@
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "BlockSettingsHooks",
+    "DefaultAttributeRegistryAccessor",
+    "DefaultAttributeRegistryMixin",
     "MaterialBuilderHooks",
     "MixinBlock",
     "MixinItem"
