+++ settings.gradle
@@ -20,7 +20,6 @@ include 'fabric-commands-v0'
 include 'fabric-containers-v0'
 include 'fabric-content-registries-v0'
 include 'fabric-crash-report-info-v1'
-include 'fabric-entity-attribute-api-v1'
 include 'fabric-events-interaction-v0'
 include 'fabric-events-lifecycle-v0'
 include 'fabric-item-groups-v0'
